Move store to top-level components directory
