+++ wsi_superpixel_guided_labeling/web_client/views/vue/components/ActiveLearning/ActiveLearningStats.vue
@@ -1,5 +1,5 @@
 <script>
-import { store } from './store.js';
+import { store } from '../store.js';
 
 export default {
     data() {
